Moves venafi addon into the framework addons package
