+++ test/e2e/framework/addon/venafi/BUILD.bazel
@@ -7,7 +7,7 @@ go_library(
         "doc.go",
         "tpp.go",
     ],
-    importpath = "github.com/jetstack/cert-manager/test/e2e/suite/issuers/venafi/addon",
+    importpath = "github.com/jetstack/cert-manager/test/e2e/framework/addon/venafi",
     tags = ["manual"],
     visibility = ["//visibility:public"],
     deps = [
