+++ test/e2e/suite/conformance/certificates/venaficloud/BUILD.bazel
@@ -9,10 +9,10 @@ go_library(
         "//pkg/apis/certmanager/v1:go_default_library",
         "//pkg/apis/meta/v1:go_default_library",
         "//test/e2e/framework:go_default_library",
+        "//test/e2e/framework/addon/venafi:go_default_library",
         "//test/e2e/framework/helper/featureset:go_default_library",
         "//test/e2e/framework/util/errors:go_default_library",
         "//test/e2e/suite/conformance/certificates:go_default_library",
-        "//test/e2e/suite/issuers/venafi/addon:go_default_library",
         "@com_github_onsi_ginkgo//:go_default_library",
         "@com_github_onsi_gomega//:go_default_library",
         "@io_k8s_apimachinery//pkg/apis/meta/v1:go_default_library",
