+++ test/e2e/suite/issuers/venafi/BUILD.bazel
@@ -19,7 +19,6 @@ filegroup(
     name = "all-srcs",
     srcs = [
         ":package-srcs",
-        "//test/e2e/suite/issuers/venafi/addon:all-srcs",
         "//test/e2e/suite/issuers/venafi/tpp:all-srcs",
     ],
     tags = ["automanaged"],
